added checks to delete team helper functions

+++ models/light_maps.js
@@ -97,6 +97,10 @@ module.exports.deleteMap = async function(mapId) {
 }
 
 module.exports.projectCleanup = async function(projectId) {
+    
+    if (!await Maps.find({project: projectId})){
+        return
+    }
     return await Maps.deleteMany({ project: projectId })
 }
