+++ models/moving_maps.js
@@ -104,6 +104,11 @@ module.exports.deleteMap = async function(mapId) {
 }
 
 module.exports.projectCleanup = async function(projectId) {
+
+    if (!await Maps.find({project: projectId})){
+        return
+    }
+
     return await Maps.deleteMany({ project: projectId })
 }
