+++ models/stationary_maps.js
@@ -115,6 +115,9 @@ module.exports.deleteMap = async function(mapId) {
 }
 
 module.exports.projectCleanup = async function(projectId) {
+    if (!await Maps.find({project: projectId})){
+        return
+    }
     return await Maps.deleteMany({ project: projectId })
 }
